char(main): Add level 1 stats
Level in Wizard, abilities based on given 10, 12, 16 (DM rules), and
rolled 13, 13, and 17.

The format of the files is not final. Only stats that are manually
tracked should be added here, as anything else should be automantically
calculated using Pilgrim. Class levels, HP, ability scores, equipment,
etc. will be tracked here.