char(main): Change race to Human
Humans gain +1 feat, which we'll need for our build. We don't get the
extra points into Charisma, but as long as we get to 19+ that's all that
matters anyway. We lose a number of other things (that I haven't tracked
yet anyway), but we also gain an extra skill point each level as well.