char(spells): Change opposition schools
We're only adding 3 levels into the Wizard class, so we're not going to
be doing a lot of damage-based spells. Therefore, we can safely add
Evocation as one of our opposition schools. Enchantment still means
little to us, so we will swap Evocation with Abjuration. We're likely
going to want some buff spells for our undead, anyway.